Fix functions appearing in variables `__all__` but not in docs for NX2.7 (#5289)
* Adjust functions appearing in `__all__` but not in docs

* clean up coloring: merge two modules make interchange private

* fix duplicate name. Probably should be changed

* fix "see also" doc of recursive_simple_cycles

* Rm internal uses of deprecated .

* Fixup warnings filters regex.

* clean up a bit more, make Node & AdjList private classes
