+++ doc/reference/algorithms/cycles.rst
@@ -8,5 +8,6 @@ Cycles
 
    cycle_basis
    simple_cycles
+   recursive_simple_cycles
    find_cycle
    minimum_cycle_basis
