+++ networkx/algorithms/tests/test_core.py
@@ -50,28 +50,24 @@ class TestCore:
     def test_trivial(self):
         """Empty graph"""
         G = nx.Graph()
-        assert nx.find_cores(G) == {}
+        assert nx.core_number(G) == {}
 
-    def test_find_cores(self):
-        core = nx.find_cores(self.G)
+    def test_core_number(self):
+        core = nx.core_number(self.G)
         nodes_by_core = [sorted(n for n in core if core[n] == val) for val in range(4)]
         assert nodes_equal(nodes_by_core[0], [21])
         assert nodes_equal(nodes_by_core[1], [17, 18, 19, 20])
         assert nodes_equal(nodes_by_core[2], [9, 10, 11, 12, 13, 14, 15, 16])
         assert nodes_equal(nodes_by_core[3], [1, 2, 3, 4, 5, 6, 7, 8])
 
-    def test_core_number(self):
-        # smoke test real name
-        cores = nx.core_number(self.G)
-
-    def test_find_cores2(self):
-        core = nx.find_cores(self.H)
+    def test_core_number2(self):
+        core = nx.core_number(self.H)
         nodes_by_core = [sorted(n for n in core if core[n] == val) for val in range(3)]
         assert nodes_equal(nodes_by_core[0], [0])
         assert nodes_equal(nodes_by_core[1], [1, 3])
         assert nodes_equal(nodes_by_core[2], [2, 4, 5, 6])
 
-    def test_directed_find_cores(self):
+    def test_directed_core_number(self):
         """core number had a bug for directed graphs found in issue #1959"""
         # small example where too timid edge removal can make cn[2] = 3
         G = nx.DiGraph()
